+++ src/complex/mod.rs
@@ -1,0 +1,5 @@
+use num_complex::Complex;
+
+pub type C64 = Complex<f64>;
+
+pub mod vector;