+++ src/complex/vector.rs
@@ -3,16 +3,14 @@ use crate::traits::fp::FPVector;
 use crate::traits::math::{Vector, Normed, Norm, InnerProduct};
 use crate::traits::sugar::VecOps;
 
-pub type C64 = Complex<f64>;
-
 impl Vector for Complex<f64> {
     type Scalar = Self;
 
-    fn add_vec<'a, 'b>(&'a self, rhs: &'b Self) -> Self {
+    fn add_vec(&self, rhs: &Self) -> Self {
         self + rhs
     }
 
-    fn sub_vec<'a, 'b>(&'a self, rhs: &'b Self) -> Self {
+    fn sub_vec(&self, rhs: &Self) -> Self {
         self - rhs
     }
 
@@ -70,15 +68,15 @@ impl FPVector for Vec<Complex<f64>> {
     fn filter<F>(&self, f: F) -> Self
     where
             F: Fn(Self::Scalar) -> bool {
-        self.into_iter().filter(|&x| f(*x)).map(|&t| t).collect()
+        self.iter().filter(|&x| f(*x)).cloned().collect()
     }
 
     fn take(&self, n: usize) -> Self {
-        self.iter().take(n).map(|&x| x).collect()
+        self.iter().take(n).cloned().collect()
     }
 
     fn skip(&self, n: usize) -> Self {
-        self.iter().skip(n).map(|&x| x).collect()
+        self.iter().skip(n).cloned().collect()
     }
 
     fn sum(&self) -> Self::Scalar {
@@ -93,11 +91,11 @@ impl FPVector for Vec<Complex<f64>> {
 impl Vector for Vec<Complex<f64>> {
     type Scalar = Complex<f64>;
 
-    fn add_vec<'a, 'b>(&'a self, rhs: &'b Self) -> Self {
+    fn add_vec(&self, rhs: &Self) -> Self {
         self.zip_with(|x, y| x + y, rhs)
     }
 
-    fn sub_vec<'a, 'b>(&'a self, rhs: &'b Self) -> Self {
+    fn sub_vec(&self, rhs: &Self) -> Self {
         self.zip_with(|x, y| x - y, rhs)
     }